--- src/components/Header/Header.scss
@@ -1,49 +1,0 @@
-@use "@styles/_variables.scss" as *;
-
-.header {
-  background-color: $primary-100;
-  color: $white;
-  padding: 20px;
-  display: flex;
-  justify-content: space-between;
-
-  &__title {
-    font-size: 1.5rem;
-    font-weight: bold;
-
-    a {
-      text-decoration: none;
-      color: $white;
-
-      &:focus,
-      &:active {
-        color: $white; // 클릭 시 색상이 변하지 않도록 설정
-        outline: none; // 클릭 시 기본 아웃라인 제거
-      }
-    }
-  }
-
-  &__nav {
-    display: flex;
-    justify-content: space-evenly;
-
-    &-link {
-      display: flex;
-      gap: 15px;
-
-      a {
-        text-decoration: none;
-        color: $white;
-        font-size: 1rem;
-        font-weight: 500;
-        padding: 5px 10px;
-
-        &:focus,
-        &:active {
-          color: $white; // 클릭 시 색상이 변하지 않도록 설정
-          outline: none; // 클릭 시 기본 아웃라인 제거
-        }
-      }
-    }
-  }
-}
