--- src/components/RightSidebar/RightSidebar.scss
@@ -1,30 +1,0 @@
-.right-sidebar {
-  &__section {
-    margin-bottom: 20px;
-
-    &-title {
-      font-size: 18px;
-      margin-bottom: 10px;
-      color: #333;
-    }
-
-    &-content {
-      font-size: 14px;
-      color: #555;
-
-      ul {
-        list-style: none;
-        padding: 0;
-
-        li {
-          margin: 8px 0;
-
-          &:hover {
-            text-decoration: underline;
-            cursor: pointer;
-          }
-        }
-      }
-    }
-  }
-}
