+++ src/main.jsx
@@ -1,5 +1,4 @@
 import { createRoot } from "react-dom/client";
-import "./index.scss";
-import App from "./App.jsx";
+import App from "./core/App";
 
 createRoot(document.getElementById("root")).render(<App />);
