+++ src/pages/Home.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { Helmet } from "react-helmet-async"; // Helmet 추가
-import PostList from "../../components/PostList/PostList";
-import { getPosts } from "../../utils/postService";
+import PostList from "../shared/components/PostList/PostList";
+import { getPosts } from "../features/Post/services/postService";
 import "./Home.scss";
 
 function Home() {
