+++ src/pages/Home.scss
