+++ src/shared/components/Footer/Footer.scss
@@ -1,14 +1,13 @@
 @use "@styles/_variables.scss" as *;
 
 .footer {
-  background-color: $primary-100;
-  color: $white;
+  color: $black;
   padding: 20px;
   text-align: center;
-
+  border-top: 0.5px dotted $primary-100;
   &__text {
     font-size: 1rem;
-    color: $white;
+    color: $black;
   }
 
   &__links {
