+++ src/shared/components/Header/Header.module.scss
@@ -1,0 +1,59 @@
+@use "@styles/_variables.scss" as *;
+
+.header {
+  color: $black;
+  padding: 10px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border-bottom: 0.5px dotted $primary-100;
+
+  &__wrapper {
+    max-width: 1100px;
+    min-width: 961px;
+    display: flex;
+    justify-content: space-between;
+  }
+  &__logo {
+    width: 40px;
+    height: 40px;
+  }
+  &__title {
+    font-size: 1.5rem;
+    font-weight: bold;
+
+    a {
+      text-decoration: none;
+      color: $black;
+
+      &:focus,
+      &:active {
+        color: $black;
+        outline: none;
+      }
+    }
+  }
+
+  &__nav {
+    display: flex;
+    justify-content: space-evenly;
+    gap: 15px;
+
+    &_link {
+      display: flex;
+
+      a {
+        text-decoration: none;
+        color: $black;
+        font-size: 1rem;
+        font-weight: 500;
+        line-height: 45px;
+        &:focus,
+        &:active {
+          color: $black;
+          outline: none;
+        }
+      }
+    }
+  }
+}
