+++ src/shared/components/LeftSidebar/LeftSidebar.scss
