+++ src/shared/components/PostItem/PostItem.jsx
@@ -1,29 +1,14 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import { Helmet } from "react-helmet-async";
+import "./PostItem.scss";
 
-import "./Post.scss";
-
-function Post({ post }) {
+function PostItem({ post }) {
   const handleTagClick = (e) => {
     e.stopPropagation();
   };
 
   return (
     <>
-      <Helmet>
-        <title>{post.title} - My Blog</title>
-        <meta name="description" content={post.excerpt} />
-        <meta property="og:title" content={post.title} />
-        <meta property="og:description" content={post.excerpt} />
-        <meta property="og:image" content={post.thumbnail} />
-        <meta
-          property="og:url"
-          content={`https://yourblog.com/post/${post.id}`}
-        />
-        <meta name="twitter:card" content="summary_large_image" />
-      </Helmet>
-
       <Link to={`/post/${post.id}`} className="post">
         <div className="post__content">
           <div className="post__info">
@@ -44,4 +29,4 @@ function Post({ post }) {
   );
 }
 
-export default Post;
+export default PostItem;
