+++ src/shared/components/PostItem/PostItem.scss
