+++ src/shared/components/PostList/PostList.scss
