+++ src/shared/components/Tag/Tag.scss
