+++ vite.config.js
@@ -6,7 +6,7 @@ export default defineConfig({
   plugins: [react()],
   resolve: {
     alias: {
-      "@styles": path.resolve(__dirname, "./src/styles"),
+      "@styles": path.resolve(__dirname, "./src/shared/styles"),
       "@components": path.resolve(__dirname, "./src/components"),
       "@pages": path.resolve(__dirname, "./src/pages"),
     },
